+++ src/settings.rs
@@ -6,7 +6,7 @@ pub const BG_CHAR: char = '-';
 pub const SHIP: char = '■';
 pub const HIT: char = '#';
 pub const STUNK: char = 'X';
-pub const TABLE_SIZE: usize = 10;
+pub const TABLE_SIZE: i32 = 10;
 pub const TITLE: &str = "Battleship-rs";
 
 //
